reorganize auth folders
