+++ src/auth/auth.module.ts
@@ -1,14 +1,13 @@
 import { Module } from '@nestjs/common';
+import { ConfigModule, ConfigService } from '@nestjs/config';
 import { APP_GUARD } from '@nestjs/core';
 import { JwtModule } from '@nestjs/jwt';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
-import { UsersModule } from '@/users/users.module';
+import { AuthService } from './auth.service';
+import { JwtAuthGuard } from './guards';
+import { LocalStrategy, JwtStrategy } from './strategies';
 import { PasswordModule } from '@/password/password.module';
-import { LocalStrategy } from './local.strategy';
-import { JwtStrategy } from './jwt.strategy';
-import { JwtAuthGuard } from './jwt-auth.guard';
+import { UsersModule } from '@/users/users.module';
 
 @Module({
   imports: [
