+++ src/auth/guards/index.ts
@@ -1,0 +1,2 @@
+export * from './local-auth.guard';
+export * from './jwt-auth.guard';
