+++ src/auth/strategies/index.ts
@@ -1,0 +1,2 @@
+export * from './local.strategy';
+export * from './jwt.strategy';
