+++ src/auth/strategies/local.strategy.ts
@@ -1,7 +1,7 @@
-import { Strategy } from 'passport-local';
-import { PassportStrategy } from '@nestjs/passport';
 import { Injectable, UnauthorizedException } from '@nestjs/common';
-import { AuthService } from './auth.service';
+import { PassportStrategy } from '@nestjs/passport';
+import { Strategy } from 'passport-local';
+import { AuthService } from '../auth.service';
 import { User } from '@/users/entities/user.entity';
 
 @Injectable()
